Rename class to follow industry terms
